+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,0 +1,19 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminController {
+    @GetMapping
+    public String adminPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationAdminPage() {
+        return "admin/reservation-legacy";
+    }
+}
