+++ src/main/java/roomescape/controller/HomeController.java
@@ -1,0 +1,12 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+    @GetMapping("/")
+    public String redirectReservationPage() {
+        return "redirect:/admin/reservation";
+    }
+}
